+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerWithTimezoneDst.test.jsx
@@ -2,12 +2,10 @@ import React from 'react';
 import DateTimePickerTimezoneExampleTemplate from '../../common/DateTimePickerTimezoneExampleTemplate';
 
 const DateTimePickerExample = () => (
-  <>
-    <DateTimePickerTimezoneExampleTemplate
-      value="2017-11-05T01:30:00"
-      timeZone="America/Chicago"
-    />
-  </>
+  <DateTimePickerTimezoneExampleTemplate
+    value="2017-11-05T01:30:00"
+    timeZone="America/Chicago"
+  />
 );
 
 export default DateTimePickerExample;