+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerWithTimezoneMissingHour.test.jsx
@@ -2,11 +2,9 @@ import React from 'react';
 import DateTimePickerTimezoneExampleTemplate from '../../common/DateTimePickerTimezoneExampleTemplate';
 
 const DateTimePickerExample = () => (
-  <div>
-    <DateTimePickerTimezoneExampleTemplate
-      timeZone="America/New_York"
-    />
-  </div>
+  <DateTimePickerTimezoneExampleTemplate
+    timeZone="America/New_York"
+  />
 );
 
 export default DateTimePickerExample;