+++ packages/terra-date-time-picker/tests/jest/TimeClarification.test.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 /* eslint-disable import/no-extraneous-dependencies */
 import { shallowWithIntl } from 'terra-enzyme-intl';
-import moment from 'moment-timezone';
 import TimeClarification from '../../lib/_TimeClarification';
 
 it('should render a default date time picker', () => {
@@ -13,6 +12,8 @@ it('should render a default date time picker', () => {
       onStandardTimeButtonClick={jest.fn()}
       onOffsetButtonClick={jest.fn()}
       onRequestClose={jest.fn()}
+      timeZone="America/Chicago"
+
     />
   ));
   const wrapper = datePicker.dive();
@@ -29,6 +30,7 @@ it('should render a disabled time clarification', () => {
       onOffsetButtonClick={jest.fn()}
       onRequestClose={jest.fn()}
       disabled
+      timeZone="America/Chicago"
     />
   ));
   const wrapper = datePicker.dive();
@@ -36,7 +38,6 @@ it('should render a disabled time clarification', () => {
 });
 
 it('should render offset button after daylight savings button clicked', () => {
-  moment.tz.guess = jest.fn(() => 'America/Chicago');
   const datePicker = shallowWithIntl((
     <TimeClarification
       ambiguousDateTime="2019-08-15T10:30:00"
@@ -47,6 +48,7 @@ it('should render offset button after daylight savings button clicked', () => {
       onOffsetButtonClick={jest.fn()}
       onRequestClose={jest.fn()}
       disabled
+      timeZone="America/Chicago"
     />
   ));
   const wrapper = datePicker.dive();
@@ -55,7 +57,6 @@ it('should render offset button after daylight savings button clicked', () => {
 });
 
 it('should render offset button after standard time button clicked', () => {
-  moment.tz.guess = jest.fn(() => 'America/Chicago');
   const datePicker = shallowWithIntl((
     <TimeClarification
       ambiguousDateTime="2019-02-15T10:30:00"
@@ -66,6 +67,7 @@ it('should render offset button after standard time button clicked', () => {
       onOffsetButtonClick={jest.fn()}
       onRequestClose={jest.fn()}
       disabled
+      timeZone="America/Chicago"
     />
   ));
   const wrapper = datePicker.dive();
@@ -74,7 +76,6 @@ it('should render offset button after standard time button clicked', () => {
 });
 
 it('should render offset button after daylight savings button clicked in the southern hemisphere timezone', () => {
-  moment.tz.guess = jest.fn(() => 'Australia/Sydney');
   const datePicker = shallowWithIntl((
     <TimeClarification
       ambiguousDateTime="2019-04-01T02:30:00"
@@ -85,6 +86,7 @@ it('should render offset button after daylight savings button clicked in the sou
       onOffsetButtonClick={jest.fn()}
       onRequestClose={jest.fn()}
       disabled
+      timeZone="Australia/Sydney"
     />
   ));
   const wrapper = datePicker.dive();
@@ -93,7 +95,6 @@ it('should render offset button after daylight savings button clicked in the sou
 });
 
 it('should render offset button after standard time button clicked in the southern hemisphere timezone', () => {
-  moment.tz.guess = jest.fn(() => 'Australia/Sydney');
   const datePicker = shallowWithIntl((
     <TimeClarification
       ambiguousDateTime="2019-10-01T02:30:00"
@@ -104,6 +105,7 @@ it('should render offset button after standard time button clicked in the southe
       onOffsetButtonClick={jest.fn()}
       onRequestClose={jest.fn()}
       disabled
+      timeZone="Australia/Sydney"
     />
   ));
   const wrapper = datePicker.dive();