+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -1,10 +1,10 @@
 ChangeLog
 =========
-### Added
-* Added new `timeZone` prop which takes a valid timezone string. This enables date-time-picker set to a particular timezone. The default timezone  is set to the local/browser timezone.
 
 Unreleased
 ----------
+### Added
+* Added new `timeZone` prop which takes a valid timezone string. This enables date-time-picker set to a particular timezone. The default timezone  is set to the local/browser timezone.
 
 4.42.0 - (July 28, 2020)
 ------------------